major refactoring, maing state access to db easier

+++ packages/states/draft/confirmdraftdeletionstate.py
@@ -7,26 +7,33 @@ __author__ = "aneanet"
 class ConfirmDraftDeletionState(SelectDraftUpdateState):
     """
     Concrete state implementation.
+
     Lets the user confirm or abort deletion of a previously chosen draft.
     """
 
     @property
     def welcome_message(self):
         message = "It seems the draft you selected no longer exists..."
-        user_drafts = self.context.get_posts(post_id=self.post_id)
-        if len(user_drafts) > 0:
-            post_title = user_drafts[0]["title"]
-            message = "Do you <b>really</b> want to <b>delete</b> draft <b>" + post_title + "</b>?"
+
+        post = self.context.a_get_post(self.post_id)
+        if post is not None:
+            message = "Do you <b>really</b> want to <b>delete</b> draft <b>" + post.title + "</b>?"
+
         return message
 
     @property
     def callback_options(self):
+
+        # add button to return to draft list
         reply_options = [{"text": "<< drafts", "callback_data": "/deletedraft"}, []]
 
-        user_drafts = self.context.get_posts(post_id=self.post_id)
-        if len(user_drafts) > 0:
+        # if post still exists, show buttn to confirm final deletion
+        post = self.context.a_get_post(self.post_id)
+        if post is not None:
             reply_options.append({"text": "YES, delete", "callback_data": "/confirmdraftdeletion /confirm"})
             reply_options.append([])
+
+        # add button to return to main menu
         reply_options.append({"text": "<< main menu", "callback_data": "/mainmenu"})
 
         return reply_options
@@ -35,36 +42,34 @@ class ConfirmDraftDeletionState(SelectDraftUpdateState):
         next_state = self
         command_array = data.split(" ")
 
-        # only accept "/confirmdraftdeletion ..." callback queries, have super() handle everything else
-        if len(command_array) > 1 and command_array[0] == "/confirmdraftdeletion":
+        # only accept "/confirmdraftdeletion <confirm/back>" callback queries
+        if len(command_array) == 2 and command_array[0] == "/confirmdraftdeletion":
 
-            # deletion of draft confirmed or aborted - /confirmdraftdeletion <confirm/back>
-            if len(command_array) == 2:
+            # confirmed draft deletion
+            if command_array[1] == "/confirm":
 
-                # confirmed draft deletion
-                if command_array[1] == "/confirm":
+                deleted_post = self.context.a_delete_post(self.post_id)
 
-                    user_drafts = self.context.get_posts(post_id=self.post_id)
-                    if len(user_drafts) > 0:
-                        post_title = user_drafts[0]["title"]
-                        self.context.delete_post(self.post_id)
-                        self.context.edit_message_text(chat_id, message_id
-                                                             , "Draft <b>" + post_title + "</b> has been <b>deleted</b>."
-                                                             , parse_mode=ParseMode.HTML.value)
+                # post removal successful
+                if deleted_post is not None:
+                    self.context.edit_message_text(chat_id, message_id
+                                                   , "Draft <b>" + deleted_post.title + "</b> has been <b>deleted</b>."
+                                                   , parse_mode=ParseMode.HTML.value)
+                # post removal not successful
+                else:
+                    self.context.edit_message_text(chat_id, self.message_id
+                                              , "It seems the draft you selected no longer exists..."
+                                              , parse_mode=ParseMode.HTML.value)
 
-                    else:
-                        self.context.edit_message_text(chat_id, self.message_id
-                                                  , "It seems the draft you selected no longer exists..."
-                                                  , parse_mode=ParseMode.HTML.value)
-
-                    # show remaining drafts for deletion
-                    if len(self.context.get_posts(user_id=user_id, status="draft")) > 0:
-                        from packages.states.draft.deletedraftstate import DeleteDraftState
-                        next_state = DeleteDraftState(self.context, user_id, chat_id=chat_id)
-                    # no remaining drafts -> automatically go back to main menu
-                    else:
-                        from packages.states.navigation.idlestate import IdleState
-                        next_state = IdleState(self.context, user_id, chat_id=chat_id)
+                # show remaining drafts for deletion
+                post_tags = self.context.a_get_post_tags(self.post_id)
+                if len(post_tags) > 0:
+                    from packages.states.draft.deletedraftstate import DeleteDraftState
+                    next_state = DeleteDraftState(self.context, user_id, chat_id=chat_id)
+                # no remaining drafts -> automatically go back to main menu
+                else:
+                    from packages.states.navigation.idlestate import IdleState
+                    next_state = IdleState(self.context, user_id, chat_id=chat_id)
 
         else:
             next_state = super().process_callback_query(user_id, chat_id, message_id, data)
